+++ src/pot.sol
@@ -43,32 +43,39 @@ pragma solidity ^0.8.20;
 */
 
 interface VatLike {
-    function move(address,address,uint256) external;
-    function suck(address,address,uint256) external;
+    function move(address, address, uint256) external;
+    function suck(address, address, uint256) external;
 }
 
 contract Pot {
     // --- Auth ---
-    mapping (address => uint) public wards;
-    function rely(address guy) external auth { wards[guy] = 1; }
-    function deny(address guy) external auth { wards[guy] = 0; }
-    modifier auth {
+    mapping(address => uint256) public wards;
+
+    function rely(address guy) external auth {
+        wards[guy] = 1;
+    }
+
+    function deny(address guy) external auth {
+        wards[guy] = 0;
+    }
+
+    modifier auth() {
         require(wards[msg.sender] == 1, "Pot/not-authorized");
         _;
     }
 
     // --- Data ---
-    mapping (address => uint256) public pie;  // Normalised Savings Dai [wad]
+    mapping(address => uint256) public pie; // Normalised Savings Dai [wad]
 
-    uint256 public Pie;   // Total Normalised Savings Dai  [wad]
-    uint256 public dsr;   // The Dai Savings Rate          [ray]
-    uint256 public chi;   // The Rate Accumulator          [ray]
+    uint256 public Pie; // Total Normalised Savings Dai  [wad]
+    uint256 public dsr; // The Dai Savings Rate          [ray]
+    uint256 public chi; // The Rate Accumulator          [ray]
 
-    VatLike public vat;   // CDP Engine
-    address public vow;   // Debt Engine
-    uint256 public rho;   // Time of last drip     [unix epoch time]
+    VatLike public vat; // CDP Engine
+    address public vow; // Debt Engine
+    uint256 public rho; // Time of last drip     [unix epoch time]
 
-    uint256 public live;  // Active Flag
+    uint256 public live; // Active Flag
 
     // --- Init ---
     constructor(address vat_) {
@@ -82,23 +89,31 @@ contract Pot {
 
     // --- Math ---
     uint256 constant ONE = 10 ** 27;
-    function _rpow(uint x, uint n, uint base) internal pure returns (uint z) {
+
+    function _rpow(uint256 x, uint256 n, uint256 base) internal pure returns (uint256 z) {
         assembly {
-            switch x case 0 {switch n case 0 {z := base} default {z := 0}}
+            switch x
+            case 0 {
+                switch n
+                case 0 { z := base }
+                default { z := 0 }
+            }
             default {
-                switch mod(n, 2) case 0 { z := base } default { z := x }
-                let half := div(base, 2)  // for rounding.
-                for { n := div(n, 2) } n { n := div(n,2) } {
+                switch mod(n, 2)
+                case 0 { z := base }
+                default { z := x }
+                let half := div(base, 2) // for rounding.
+                for { n := div(n, 2) } n { n := div(n, 2) } {
                     let xx := mul(x, x)
-                    if iszero(eq(div(xx, x), x)) { revert(0,0) }
+                    if iszero(eq(div(xx, x), x)) { revert(0, 0) }
                     let xxRound := add(xx, half)
-                    if lt(xxRound, xx) { revert(0,0) }
+                    if lt(xxRound, xx) { revert(0, 0) }
                     x := div(xxRound, base)
-                    if mod(n,2) {
+                    if mod(n, 2) {
                         let zx := mul(z, x)
-                        if and(iszero(iszero(x)), iszero(eq(div(zx, x), z))) { revert(0,0) }
+                        if and(iszero(iszero(x)), iszero(eq(div(zx, x), z))) { revert(0, 0) }
                         let zxRound := add(zx, half)
-                        if lt(zxRound, zx) { revert(0,0) }
+                        if lt(zxRound, zx) { revert(0, 0) }
                         z := div(zxRound, base)
                     }
                 }
@@ -106,19 +121,19 @@ contract Pot {
         }
     }
 
-    function _rmul(uint x, uint y) internal pure returns (uint z) {
+    function _rmul(uint256 x, uint256 y) internal pure returns (uint256 z) {
         z = _mul(x, y) / ONE;
     }
 
-    function _add(uint x, uint y) internal pure returns (uint z) {
+    function _add(uint256 x, uint256 y) internal pure returns (uint256 z) {
         require((z = x + y) >= x);
     }
 
-    function _sub(uint x, uint y) internal pure returns (uint z) {
+    function _sub(uint256 x, uint256 y) internal pure returns (uint256 z) {
         require((z = x - y) <= x);
     }
 
-    function _mul(uint x, uint y) internal pure returns (uint z) {
+    function _mul(uint256 x, uint256 y) internal pure returns (uint256 z) {
         require(y == 0 || (z = x * y) / y == x);
     }
 
@@ -141,26 +156,26 @@ contract Pot {
     }
 
     // --- Savings Rate Accumulation ---
-    function drip() external returns (uint tmp) {
+    function drip() external returns (uint256 tmp) {
         require(block.timestamp >= rho, "Pot/invalid-block.timestamp");
         tmp = _rmul(_rpow(dsr, block.timestamp - rho, ONE), chi);
-        uint chi_ = _sub(tmp, chi);
+        uint256 chi_ = _sub(tmp, chi);
         chi = tmp;
         rho = block.timestamp;
         vat.suck(address(vow), address(this), _mul(Pie, chi_));
     }
 
     // --- Savings Dai Management ---
-    function join(uint wad) external {
+    function join(uint256 wad) external {
         require(block.timestamp == rho, "Pot/rho-not-updated");
         pie[msg.sender] = _add(pie[msg.sender], wad);
-        Pie             = _add(Pie,             wad);
+        Pie = _add(Pie, wad);
         vat.move(msg.sender, address(this), _mul(chi, wad));
     }
 
-    function exit(uint wad) external {
+    function exit(uint256 wad) external {
         pie[msg.sender] = _sub(pie[msg.sender], wad);
-        Pie             = _sub(Pie,             wad);
+        Pie = _sub(Pie, wad);
         vat.move(address(this), msg.sender, _mul(chi, wad));
     }
 }
