+++ src/spot.sol
@@ -22,7 +22,7 @@ pragma solidity ^0.8.20;
 // New deployments of this contract will need to include custom events (TO DO).
 
 interface VatLike {
-    function file(bytes32, bytes32, uint) external;
+    function file(bytes32, bytes32, uint256) external;
 }
 
 interface PipLike {
@@ -31,33 +31,38 @@ interface PipLike {
 
 contract Spotter {
     // --- Auth ---
-    mapping (address => uint) public wards;
-    function rely(address guy) external auth { wards[guy] = 1;  }
-    function deny(address guy) external auth { wards[guy] = 0; }
-    modifier auth {
+    mapping(address => uint256) public wards;
+
+    function rely(address guy) external auth {
+        wards[guy] = 1;
+    }
+
+    function deny(address guy) external auth {
+        wards[guy] = 0;
+    }
+
+    modifier auth() {
         require(wards[msg.sender] == 1, "Spotter/not-authorized");
         _;
     }
 
     // --- Data ---
     struct Ilk {
-        PipLike pip;  // Price Feed
-        uint256 mat;  // Liquidation ratio [ray]
+        PipLike pip; // Price Feed
+        uint256 mat; // Liquidation ratio [ray]
     }
 
-    mapping (bytes32 => Ilk) public ilks;
+    mapping(bytes32 => Ilk) public ilks;
 
-    VatLike public vat;  // CDP Engine
-    uint256 public par;  // ref per dai [ray]
+    VatLike public vat; // CDP Engine
+    uint256 public par; // ref per dai [ray]
 
     uint256 public live;
 
     // --- Events ---
-    event Poke(
-      bytes32 ilk,
-      bytes32 val,  // [wad]
-      uint256 spot  // [ray]
-    );
+    event Poke( // [wad]
+        // [ray]
+    bytes32 ilk, bytes32 val, uint256 spot);
 
     // --- Init ---
     constructor(address vat_) {
@@ -68,12 +73,13 @@ contract Spotter {
     }
 
     // --- Math ---
-    uint constant ONE = 10 ** 27;
+    uint256 constant ONE = 10 ** 27;
 
-    function mul(uint x, uint y) internal pure returns (uint z) {
+    function mul(uint256 x, uint256 y) internal pure returns (uint256 z) {
         require(y == 0 || (z = x * y) / y == x);
     }
-    function rdiv(uint x, uint y) internal pure returns (uint z) {
+
+    function rdiv(uint256 x, uint256 y) internal pure returns (uint256 z) {
         z = mul(x, ONE) / y;
     }
 
@@ -83,12 +89,14 @@ contract Spotter {
         if (what == "pip") ilks[ilk].pip = PipLike(pip_);
         else revert("Spotter/file-unrecognized-param");
     }
-    function file(bytes32 what, uint data) external auth {
+
+    function file(bytes32 what, uint256 data) external auth {
         require(live == 1, "Spotter/not-live");
         if (what == "par") par = data;
         else revert("Spotter/file-unrecognized-param");
     }
-    function file(bytes32 ilk, bytes32 what, uint data) external auth {
+
+    function file(bytes32 ilk, bytes32 what, uint256 data) external auth {
         require(live == 1, "Spotter/not-live");
         if (what == "mat") ilks[ilk].mat = data;
         else revert("Spotter/file-unrecognized-param");
@@ -97,7 +105,7 @@ contract Spotter {
     // --- Update value ---
     function poke(bytes32 ilk) external {
         (bytes32 val, bool has) = ilks[ilk].pip.peek();
-        uint256 spot = has ? rdiv(rdiv(mul(uint(val), 10 ** 9), par), ilks[ilk].mat) : 0;
+        uint256 spot = has ? rdiv(rdiv(mul(uint256(val), 10 ** 9), par), ilks[ilk].mat) : 0;
         vat.file(ilk, "spot", spot);
         emit Poke(ilk, val, spot);
     }
