+++ test/flop.t.sol
@@ -17,11 +17,10 @@
 
 pragma solidity ^0.8.20;
 
-import {DSTest}  from "ds-test/test.sol";
+import {DSTest} from "ds-test/test.sol";
 import {DSToken} from "ds-token/token.sol";
 import "../src/flop.sol";
 import "../src/vat.sol";
-
 
 interface Hevm {
     function warp(uint256) external;
@@ -29,61 +28,67 @@ interface Hevm {
 
 contract Guy {
     Flopper flop;
+
     constructor(Flopper flop_) {
         flop = flop_;
         Vat(address(flop.vat())).hope(address(flop));
         DSToken(address(flop.gem())).approve(address(flop));
     }
-    function dent(uint id, uint lot, uint bid) public {
+
+    function dent(uint256 id, uint256 lot, uint256 bid) public {
         flop.dent(id, lot, bid);
     }
-    function deal(uint id) public {
+
+    function deal(uint256 id) public {
         flop.deal(id);
     }
-    function try_dent(uint id, uint lot, uint bid)
-        public returns (bool ok)
-    {
+
+    function try_dent(uint256 id, uint256 lot, uint256 bid) public returns (bool ok) {
         string memory sig = "dent(uint256,uint256,uint256)";
         (ok,) = address(flop).call(abi.encodeWithSignature(sig, id, lot, bid));
     }
-    function try_deal(uint id)
-        public returns (bool ok)
-    {
+
+    function try_deal(uint256 id) public returns (bool ok) {
         string memory sig = "deal(uint256)";
         (ok,) = address(flop).call(abi.encodeWithSignature(sig, id));
     }
-    function try_tick(uint id)
-        public returns (bool ok)
-    {
+
+    function try_tick(uint256 id) public returns (bool ok) {
         string memory sig = "tick(uint256)";
         (ok,) = address(flop).call(abi.encodeWithSignature(sig, id));
     }
 }
 
 contract Gal {
-    uint public Ash;
-    function sub(uint x, uint y) internal pure returns (uint z) {
+    uint256 public Ash;
+
+    function sub(uint256 x, uint256 y) internal pure returns (uint256 z) {
         require((z = x - y) <= x);
     }
-    function kick(Flopper flop, uint lot, uint bid) external returns (uint) {
+
+    function kick(Flopper flop, uint256 lot, uint256 bid) external returns (uint256) {
         Ash += bid;
         return flop.kick(address(this), lot, bid);
     }
-    function kiss(uint rad) external {
+
+    function kiss(uint256 rad) external {
         Ash = sub(Ash, rad);
     }
+
     function cage(Flopper flop) external {
         flop.cage();
     }
 }
 
-contract Vatish is DSToken('') {
-    uint constant ONE = 10 ** 27;
+contract Vatish is DSToken("") {
+    uint256 constant ONE = 10 ** 27;
+
     function hope(address usr) public {
-         approve(usr, type(uint).max);
+        approve(usr, type(uint256).max);
     }
-    function dai(address usr) public view returns (uint) {
-         return balanceOf[usr];
+
+    function dai(address usr) public view returns (uint256) {
+        return balanceOf[usr];
     }
 }
 
@@ -91,14 +96,14 @@ contract FlopTest is DSTest {
     Hevm hevm;
 
     Flopper flop;
-    Vat     vat;
+    Vat vat;
     DSToken gem;
 
     address ali;
     address bob;
     address gal;
 
-    function kiss(uint) public pure { }  // arbitrary callback
+    function kiss(uint256) public pure {} // arbitrary callback
 
     function setUp() public {
         hevm = Hevm(0x7109709ECfa91a80626fF3989D68f67F5b1DD12D);
@@ -129,7 +134,7 @@ contract FlopTest is DSTest {
     function test_kick() public {
         assertEq(vat.dai(gal), 0);
         assertEq(gem.balanceOf(gal), 0 ether);
-        uint id = Gal(gal).kick(flop, /*lot*/ 200 ether, /*bid*/ 5000 ether);
+        uint256 id = Gal(gal).kick(flop, /*lot*/ 200 ether, /*bid*/ 5000 ether);
         // no value transferred
         assertEq(vat.dai(gal), 0);
         assertEq(gem.balanceOf(gal), 0 ether);
@@ -144,13 +149,13 @@ contract FlopTest is DSTest {
     }
 
     function test_dent() public {
-        uint id = Gal(gal).kick(flop, /*lot*/ 200 ether, /*bid*/ 10 ether);
+        uint256 id = Gal(gal).kick(flop, /*lot*/ 200 ether, /*bid*/ 10 ether);
 
         Guy(ali).dent(id, 100 ether, 10 ether);
         // bid taken from bidder
         assertEq(vat.dai(ali), 190 ether);
         // gal receives payment
-        assertEq(vat.dai(gal),  10 ether);
+        assertEq(vat.dai(gal), 10 ether);
         assertEq(Gal(gal).Ash(), 0 ether);
 
         Guy(bob).dent(id, 80 ether, 10 ether);
@@ -162,7 +167,7 @@ contract FlopTest is DSTest {
         assertEq(vat.dai(gal), 10 ether);
 
         hevm.warp(block.timestamp + 5 weeks);
-        assertEq(gem.totalSupply(),  0 ether);
+        assertEq(gem.totalSupply(), 0 ether);
         gem.setOwner(address(flop));
         Guy(bob).deal(id);
         // gems minted on demand
@@ -172,8 +177,8 @@ contract FlopTest is DSTest {
     }
 
     function test_dent_Ash_less_than_bid() public {
-        uint id = Gal(gal).kick(flop, /*lot*/ 200 ether, /*bid*/ 10 ether);
-        assertEq(vat.dai(gal),  0 ether);
+        uint256 id = Gal(gal).kick(flop, /*lot*/ 200 ether, /*bid*/ 10 ether);
+        assertEq(vat.dai(gal), 0 ether);
 
         Gal(gal).kiss(1 ether);
         assertEq(Gal(gal).Ash(), 9 ether);
@@ -182,7 +187,7 @@ contract FlopTest is DSTest {
         // bid taken from bidder
         assertEq(vat.dai(ali), 190 ether);
         // gal receives payment
-        assertEq(vat.dai(gal),   10 ether);
+        assertEq(vat.dai(gal), 10 ether);
         assertEq(Gal(gal).Ash(), 0 ether);
 
         Guy(bob).dent(id, 80 ether, 10 ether);
@@ -194,7 +199,7 @@ contract FlopTest is DSTest {
         assertEq(vat.dai(gal), 10 ether);
 
         hevm.warp(block.timestamp + 5 weeks);
-        assertEq(gem.totalSupply(),  0 ether);
+        assertEq(gem.totalSupply(), 0 ether);
         gem.setOwner(address(flop));
         Guy(bob).deal(id);
         // gems minted on demand
@@ -204,7 +209,7 @@ contract FlopTest is DSTest {
     }
 
     function test_dent_same_bidder() public {
-        uint id = Gal(gal).kick(flop, /*lot*/ 200 ether, /*bid*/ 200 ether);
+        uint256 id = Gal(gal).kick(flop, /*lot*/ 200 ether, /*bid*/ 200 ether);
 
         Guy(ali).dent(id, 100 ether, 200 ether);
         assertEq(vat.dai(ali), 0);
@@ -213,29 +218,29 @@ contract FlopTest is DSTest {
 
     function test_tick() public {
         // start an auction
-        uint id = Gal(gal).kick(flop, /*lot*/ 200 ether, /*bid*/ 10 ether);
+        uint256 id = Gal(gal).kick(flop, /*lot*/ 200 ether, /*bid*/ 10 ether);
         // check no tick
         assertTrue(!Guy(ali).try_tick(id));
         // run past the end
         hevm.warp(block.timestamp + 2 weeks);
         // check not biddable
         assertTrue(!Guy(ali).try_dent(id, 100 ether, 10 ether));
-        assertTrue( Guy(ali).try_tick(id));
+        assertTrue(Guy(ali).try_tick(id));
         // check biddable
-        (, uint _lot,,,) = flop.bids(id);
+        (, uint256 _lot,,,) = flop.bids(id);
         // tick should increase the lot by pad (50%) and restart the auction
         assertEq(_lot, 300 ether);
-        assertTrue( Guy(ali).try_dent(id, 100 ether, 10 ether));
+        assertTrue(Guy(ali).try_dent(id, 100 ether, 10 ether));
     }
 
     function test_no_deal_after_end() public {
         // if there are no bids and the auction ends, then it should not
         // be refundable to the creator. Rather, it ticks indefinitely.
-        uint id = Gal(gal).kick(flop, /*lot*/ 200 ether, /*bid*/ 10 ether);
+        uint256 id = Gal(gal).kick(flop, /*lot*/ 200 ether, /*bid*/ 10 ether);
         assertTrue(!Guy(ali).try_deal(id));
         hevm.warp(block.timestamp + 2 weeks);
         assertTrue(!Guy(ali).try_deal(id));
-        assertTrue( Guy(ali).try_tick(id));
+        assertTrue(Guy(ali).try_tick(id));
         assertTrue(!Guy(ali).try_deal(id));
     }
 
@@ -243,7 +248,7 @@ contract FlopTest is DSTest {
         // yanking the auction should refund the last bidder's dai, credit a
         // corresponding amount of sin to the caller of cage, and delete the auction.
         // in practice, gal == (caller of cage) == (vow address)
-        uint id = Gal(gal).kick(flop, /*lot*/ 200 ether, /*bid*/ 10 ether);
+        uint256 id = Gal(gal).kick(flop, /*lot*/ 200 ether, /*bid*/ 10 ether);
 
         // confrim initial state expectations
         assertEq(vat.dai(ali), 200 ether);
@@ -255,9 +260,9 @@ contract FlopTest is DSTest {
         Guy(bob).dent(id, 80 ether, 10 ether);
 
         // confirm the proper state updates have occurred
-        assertEq(vat.dai(ali), 200 ether);  // ali's dai balance is unchanged
+        assertEq(vat.dai(ali), 200 ether); // ali's dai balance is unchanged
         assertEq(vat.dai(bob), 190 ether);
-        assertEq(vat.dai(gal),  10 ether);
+        assertEq(vat.dai(gal), 10 ether);
         assertEq(vat.sin(address(this)), 1000 ether);
 
         Gal(gal).cage(flop);
@@ -265,9 +270,9 @@ contract FlopTest is DSTest {
 
         // confirm final state
         assertEq(vat.dai(ali), 200 ether);
-        assertEq(vat.dai(bob), 200 ether);  // bob's bid has been refunded
-        assertEq(vat.dai(gal),  10 ether);
-        assertEq(vat.sin(gal),  10 ether);  // sin assigned to caller of cage()
+        assertEq(vat.dai(bob), 200 ether); // bob's bid has been refunded
+        assertEq(vat.dai(gal), 10 ether);
+        assertEq(vat.sin(gal), 10 ether); // sin assigned to caller of cage()
         (uint256 _bid, uint256 _lot, address _guy, uint48 _tic, uint48 _end) = flop.bids(id);
         assertEq(_bid, 0);
         assertEq(_lot, 0);
@@ -280,7 +285,7 @@ contract FlopTest is DSTest {
         // with no bidder to refund, yanking the auction should simply create equal
         // amounts of dai (credited to the gal) and sin (credited to the caller of cage)
         // in practice, gal == (caller of cage) == (vow address)
-        uint id = Gal(gal).kick(flop, /*lot*/ 200 ether, /*bid*/ 10 ether);
+        uint256 id = Gal(gal).kick(flop, /*lot*/ 200 ether, /*bid*/ 10 ether);
 
         // confrim initial state expectations
         assertEq(vat.dai(ali), 200 ether);
@@ -294,8 +299,8 @@ contract FlopTest is DSTest {
         // confirm final state
         assertEq(vat.dai(ali), 200 ether);
         assertEq(vat.dai(bob), 200 ether);
-        assertEq(vat.dai(gal),  10 ether);
-        assertEq(vat.sin(gal),  10 ether);  // sin assigned to caller of cage()
+        assertEq(vat.dai(gal), 10 ether);
+        assertEq(vat.sin(gal), 10 ether); // sin assigned to caller of cage()
         (uint256 _bid, uint256 _lot, address _guy, uint48 _tic, uint48 _end) = flop.bids(id);
         assertEq(_bid, 0);
         assertEq(_lot, 0);
